Moved filedlg struct to header to avoid pound including source for tests
